refactor: divide clean architecture by module
