+++ core/remote/build.gradle.kts
@@ -22,6 +22,12 @@ android {
             "MOOI_DEV_SERVER_URL",
             getLocalProperty("MOOI_DEV_SERVER_URL"),
         )
+
+        buildConfigField(
+            "String",
+            "GOOGLE_SERVER_CLIENT_ID",
+            getLocalProperty("GOOGLE_SERVER_CLIENT_ID"),
+        )
     }
 
     buildTypes {
@@ -43,4 +49,7 @@ dependencies {
     api(projects.domain)
     api(projects.data)
     api(projects.core.common)
+
+    implementation(libs.bundles.credentials)
+    implementation(libs.kakao.sdk.user)
 }
