+++ core/remote/src/main/java/com/emotionstorage/remote/di/RemoteDataSourceModule.kt
@@ -1,15 +1,21 @@
 package com.emotionstorage.remote.di
 
+import com.emotionstorage.data.dataSource.remote.AuthRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.DailyReportRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.FcmRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.GoogleRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.HomeRemoteDataSource
+import com.emotionstorage.data.dataSource.remote.KakaoRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.MyPageRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.NotificationSettingRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.TimeCapsuleRemoteDataSource
 import com.emotionstorage.data.dataSource.remote.UserRemoteDataSource
+import com.emotionstorage.remote.dataSourceImpl.AuthRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.DailyReportRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.FcmRemoteDataSourceImpl
+import com.emotionstorage.remote.dataSourceImpl.GoogleRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.HomeRemoteDataSourceImpl
+import com.emotionstorage.remote.dataSourceImpl.KakaoRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.MyPageRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.NotificationSettingRemoteDataSourceImpl
 import com.emotionstorage.remote.dataSourceImpl.TimeCapsuleRemoteDataSourceImpl
@@ -23,6 +29,20 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RemoteDataSourceModule {
+    @Binds
+    @Singleton
+    abstract fun bindAuthRemoteDataSource(authRemoteDataSourceImpl: AuthRemoteDataSourceImpl): AuthRemoteDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindKakaoRemoteDataSource(kakaoRemoteDataSourceImpl: KakaoRemoteDataSourceImpl): KakaoRemoteDataSource
+
+    @Binds
+    @Singleton
+    abstract fun bindGoogleRemoteDataSource(
+        googleRemoteDataSourceImpl: GoogleRemoteDataSourceImpl,
+    ): GoogleRemoteDataSource
+
     @Binds
     @Singleton
     abstract fun bindHomeRemoteDataSource(impl: HomeRemoteDataSourceImpl): HomeRemoteDataSource
