+++ data/src/main/java/com/emotionstorage/data/dataSource/remote/GoogleRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.auth.data.dataSource
+package com.emotionstorage.data.dataSource.remote
 
 interface GoogleRemoteDataSource {
     suspend fun getIdToken(): String
