+++ data/src/main/java/com/emotionstorage/data/dataSource/remote/KakaoRemoteDataSource.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.auth.data.dataSource
+package com.emotionstorage.data.dataSource.remote
 
 interface KakaoRemoteDataSource {
     suspend fun getIdToken(): String
