--- feat/auth/src/main/java/com/emotionstorage/auth/data/di/RepositoryModule.kt
@@ -1,17 +1,0 @@
-package com.emotionstorage.auth.data.di
-
-import com.emotionstorage.auth.data.repoImpl.AuthRepositoryImpl
-import com.emotionstorage.domain.repo.AuthRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-abstract class RepositoryModule {
-    @Binds
-    @Singleton
-    abstract fun bindAuthRepository(impl: AuthRepositoryImpl): AuthRepository
-}
