--- feat/auth/src/main/java/com/emotionstorage/auth/remote/di/ApiServiceModule.kt
@@ -1,17 +1,0 @@
-package com.emotionstorage.auth.remote.di
-
-import com.emotionstorage.auth.remote.api.AuthApiService
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import retrofit2.Retrofit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object ApiServiceModule {
-    @Singleton
-    @Provides
-    fun provideAuthApiService(retrofit: Retrofit): AuthApiService = retrofit.create(AuthApiService::class.java)
-}
